requirements-dev.txt
lark
django
pytest
sdist
build
twine
python-dotenv
click
pyyaml
